Enable legacy behavior for Npgsql breaking change with .NET 6
The timezone handling of Npgsql has changed. SensorRecorder can be fixed by correctly setting the DateTimeKind for the new sensor records, but Npgsql also complains about the timezones in ForecastLoader, which does not create DateTime objects (and those that are passed to the ForecastData points are checked against UTC), so the reason for the error is unknown.